Put logic for accounts inside service
